+++ packages/editor/src/client/services-internal/extensions/index.ts
@@ -1,3 +1,3 @@
 export * from './emojiAutocompletionSettings';
 export * from './setDataLine';
-export * from './y-rich-cursors';
+export * from './y-rich-cursors/index';
