+++ packages/editor/src/client/services-internal/extensions/y-rich-cursors/widget.spec.ts
@@ -1,14 +1,13 @@
 import { describe, expect, it } from 'vitest';
 
-import { createOffScreenIndicator, RichCaretWidget } from './y-rich-cursors';
+import { RichCaretWidget } from './widget';
 
 /**
- * Unit tests for RichCaretWidget and off-screen indicators.
+ * Unit tests for RichCaretWidget.
  *
  * Covers:
  * - Task 9.1: Updated widget DOM structure, overlay flag, sizing, isActive class
- * - Task 9.2: Off-screen indicator DOM construction and avatar fallback
- * - Requirements: 3.1, 3.2, 3.3, 3.4, 3.5, 3.10, 4.1, 4.2, 4.4
+ * - Requirements: 3.1, 3.2, 3.3, 3.4, 3.5, 3.10
  */
 
 describe('RichCaretWidget', () => {
@@ -184,103 +183,5 @@ describe('RichCaretWidget', () => {
       const widget = new RichCaretWidget('#ff0000', 'Alice', undefined, false);
       expect(widget.estimatedHeight).toBe(-1);
     });
-  });
-});
-
-describe('createOffScreenIndicator', () => {
-  it('renders an indicator with an upward arrow for direction "above"', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: '/avatar.png',
-      isActive: false,
-    });
-
-    const arrow = el.querySelector('.cm-offScreenArrow');
-    expect(arrow).not.toBeNull();
-    expect(arrow?.textContent).toBe('↑');
-  });
-
-  it('renders an indicator with a downward arrow for direction "below"', () => {
-    const el = createOffScreenIndicator({
-      direction: 'below',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: '/avatar.png',
-      isActive: false,
-    });
-
-    const arrow = el.querySelector('.cm-offScreenArrow');
-    expect(arrow?.textContent).toBe('↓');
-  });
-
-  it('renders an avatar image when imageUrlCached is provided', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: '/avatar.png',
-      isActive: false,
-    });
-
-    const img = el.querySelector(
-      'img.cm-offScreenAvatar',
-    ) as HTMLImageElement | null;
-    expect(img).not.toBeNull();
-    expect(img?.src).toContain('/avatar.png');
-  });
-
-  it('renders initials fallback when imageUrlCached is undefined', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: undefined,
-      isActive: false,
-    });
-
-    const img = el.querySelector('img.cm-offScreenAvatar');
-    expect(img).toBeNull();
-
-    const initials = el.querySelector('.cm-offScreenInitials');
-    expect(initials).not.toBeNull();
-    expect(initials?.textContent).toBe('A');
-  });
-
-  it('applies cm-yRichCursorActive class when isActive is true', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: '/avatar.png',
-      isActive: true,
-    });
-
-    expect(el.classList.contains('cm-yRichCursorActive')).toBe(true);
-  });
-
-  it('does NOT apply cm-yRichCursorActive class when isActive is false', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: '/avatar.png',
-      isActive: false,
-    });
-
-    expect(el.classList.contains('cm-yRichCursorActive')).toBe(false);
-  });
-
-  it('applies border-color from the color parameter', () => {
-    const el = createOffScreenIndicator({
-      direction: 'above',
-      color: '#ff0000',
-      name: 'Alice',
-      imageUrlCached: undefined,
-      isActive: false,
-    });
-
-    expect(el.style.borderColor).toBe('#ff0000');
   });
 });
